requirements.txt
nudenet>=3.4.2
Pillow==10.2.0
numpy==1.26.4
requests>=2.31.0 